egressgateway: refactor manager logic
This commit refactors the egress gateway manager in order to provide a
single `reconcile()` method, responsible for adding and removing entries
to and from the egress_policy map, that will be invoked on all events
received by the manager.

In addition to this, the manager will now wait for the k8s cache to be
fully synced before running its first reconciliation, in order to always
have the egress_policy map in a consistent state with the k8s
configuration.

Fixes: #17380
Fixes: #17753
Signed-off-by: Gilberto Bertin <gilberto@isovalent.com>